Make Automerge.Text<->CodeMirror code more cohesive
